[PATCH] Enable pagination on Client Connection (#917)
This pull request introduces a new upstream patch.

In a [recent upstream
release](https://github.com/auth0/terraform-provider-auth0/releases/tag/v1.22.0),
an extra GET call was implemented on Create for ConnectionClients.
This call does not implement pagination properly, so that when there are
more clients in existence than clients returned per page, it is possible
that Create can fail if the new client's ID is on page 2 of the request.
This intermittent failure was reported in #909.

There is no obvious way to test this. The default
clients-returned-per-page is 50; our Auth0 instance does not permit that
many clients.
I've tested this locally by implementing the upstream provider with a
very small PerPage for the Client, which reproduced the error reported
in #909 reliably.

This is also something that we should be able to fix upstream, so I'll
open a pull request there as well.